+++ src/modules/sbstudio/math/nearest_neighbors.py
@@ -1,6 +1,6 @@
 """Algorithm to find the nearest neighbors in a set of points."""
 
-from numpy import array, fill_diagonal, inf, newaxis, searchsorted, sum
+from numpy import array, fill_diagonal, inf, newaxis, nonzero, ptp, searchsorted, sum
 
 __all__ = ("find_nearest_neighbors",)
 
@@ -42,6 +42,19 @@ def _nearest_neighbors_brute_force(points):
     return points[first], points[second], dist_sq_matrix[first, second] ** 0.5
 
 
+def _reorder_along_principal_axis(points):
+    """Given a list of points, finds the principal axis along which the dataset
+    is the "widest" and then sorts the points in ascending order along this
+    axis.
+
+    Returns:
+        the sorted points and the index of the principal axis
+    """
+    principal_axis = ptp(points, axis=0).argmax()
+    points = points[points[:, principal_axis].argsort(), :]
+    return points, principal_axis
+
+
 def _nearest_neighbors_divide_and_conquer(points):
     """Finds the nearest neighbors in a 2D or 3D point set given as a NumPy
     array (one point per row), using an O(n log(n) log(n)) divide-and-conquer
@@ -52,8 +65,7 @@ def _nearest_neighbors_divide_and_conquer(points):
         return None, None, inf
 
     # Find the principal axis along which the dataset is the "widest"
-    principal_axis = points.ptp(axis=0).argmax()
-    points = points[points[:, principal_axis].argsort(), :]
+    points, principal_axis = _reorder_along_principal_axis(points)
     p, q, dist_sq = _nearest_neighbors_divide_and_conquer_step(points, principal_axis)
     return p, q, dist_sq**0.5
 
@@ -128,9 +140,64 @@ def find_nearest_neighbors(points):
         the coordinates of the first and the second point in the closest point
         pair, and their distance
     """
+    if not points:
+        # we have no points
+        return None, None, inf
+
     points = array(points, dtype=float)
     if points.shape[0] < 2:
         # we have less than two points
         return None, None, inf
 
     return _nearest_neighbors_divide_and_conquer(points)
+
+
+def find_all_point_pairs_closer_than(points, threshold):
+    """Finds all point pairs that are closer than the given threshold.
+
+    This is a slow brute-force algorithm, but it does not have to be executed
+    frequently so it is probably okay.
+
+    Parameters:
+        points: the input, either as a list-of-points where each point may be
+            a tuple or a list, or as a NumPy array where each row is a point
+
+    Returns:
+        the coordinate pairs representing all pairs of points that are closer
+        than the given threshold
+    """
+    result = []
+    if not points:
+        # we have no points
+        return result
+
+    points = array(points, dtype=float)
+
+    num_points, _ = points.shape
+    if num_points < 2:
+        return result
+
+    threshold_sq = threshold**2
+    points, principal_axis = _reorder_along_principal_axis(points)
+    for i, p in enumerate(points):
+        max_coord = p[principal_axis] + threshold
+        end = None
+        for j, q in enumerate(points[i + 1 :]):
+            if q[principal_axis] >= max_coord:
+                end = i + j + 1
+
+        dsq = _get_distance_sq_matrix_pairs(points[i : i + 1], points[i + 1 : end])[0]
+        qs = points[nonzero(dsq < threshold_sq)[0] + i + 1]
+
+        result.extend((p, q) for q in qs)
+
+    return result
+
+
+def test():
+    points = array([[1, 2, 3], [4, 5, 6], [7, 8, 9], [10, 11, 12]])
+    print(find_all_point_pairs_closer_than(points, 6))
+
+
+if __name__ == "__main__":
+    test()