+++ src/modules/sbstudio/plugin/views.py
@@ -1,4 +1,10 @@
-from typing import Optional
+from __future__ import annotations
+
+from typing import Iterable, Optional, TYPE_CHECKING, Tuple
+
+if TYPE_CHECKING:
+    from bpy.types import Area
+    from bpy.types import SpaceView3D
 
 from .utils import with_screen
 
@@ -11,7 +17,7 @@ __all__ = (
 
 
 @with_screen
-def find_all_3d_views(screen: Optional[str] = None):
+def find_all_3d_views(screen: Optional[str] = None) -> Iterable[SpaceView3D]:
     """Finds all 3D views in the Blender screen with the given name, and returns
     an iterator that iterates over them.
 
@@ -19,15 +25,17 @@ def find_all_3d_views(screen: Optional[str] = None):
         screen: the name of the Blender screen to scan; `None` means to use the
             current screen
 
-    Returns:
-        Optional[SpaceView3D]: the first 3D view that we find in the given
-            Blender screen, or ``None`` if no 3D view was found
+    Yields:
+        the 3D views that we find in the given Blender screen
     """
-    for space, _area in find_all_3d_views_and_their_areas(screen):
+    for space, _area in _find_all_3d_views_and_their_areas(screen):
         yield space
 
 
-def find_all_3d_views_and_their_areas(screen: Optional[str] = None):
+@with_screen
+def find_all_3d_views_and_their_areas(
+    screen: Optional[str] = None,
+) -> Iterable[Tuple[SpaceView3D, Area]]:
     """Finds all 3D views in the Blender screen with the given name, and returns
     an iterator that iterates over them and their containing areas.
 
@@ -36,10 +44,17 @@ def find_all_3d_views_and_their_areas(screen: Optional[str] = None):
             current screen
 
     Yields:
-        (Optional[SpaceView3D], Optional[Area]): all 3D views that we find in
-            the given Blender screen, and their corresponding areas
+        all 3D views that we find in the given Blender screen, and their
+        corresponding areas, in tuples
     """
-    for area in screen.areas:
+    # Now that the decorator resolved the screen name, we can pass it on
+    return _find_all_3d_views_and_their_areas(screen)
+
+
+def _find_all_3d_views_and_their_areas(
+    screen: Optional[str] = None,
+) -> Iterable[Tuple[SpaceView3D, Area]]:
+    for area in screen.areas:  # type: ignore
         if area.type == "VIEW_3D":
             for space in area.spaces:
                 if space.type == "VIEW_3D":
@@ -47,7 +62,7 @@ def find_all_3d_views_and_their_areas(screen: Optional[str] = None):
 
 
 @with_screen
-def find_one_3d_view(screen: Optional[str] = None):
+def find_one_3d_view(screen: Optional[str] = None) -> Optional[SpaceView3D]:
     """Finds a 3D view in the Blender screen with the given name, and returns
     the view itself.
 
@@ -56,14 +71,16 @@ def find_one_3d_view(screen: Optional[str] = None):
             current screen
 
     Returns:
-        Optional[SpaceView3D]: the first 3D view that we find in the given
-            Blender screen, or ``None`` if no 3D view was found
+        the first 3D view that we find in the given Blender screen, or ``None``
+        if no 3D view was found
     """
     return find_one_3d_view_and_its_area(screen)[0]
 
 
 @with_screen
-def find_one_3d_view_and_its_area(screen: Optional[str] = None):
+def find_one_3d_view_and_its_area(
+    screen: Optional[str] = None,
+) -> Tuple[Optional[SpaceView3D], Optional[Area]]:
     """Finds a 3D view in the Blender screen with the given name, and returns
     the view and its containing area in a tuple.
 
@@ -72,11 +89,10 @@ def find_one_3d_view_and_its_area(screen: Optional[str] = None):
             current screen
 
     Returns:
-        (Optional[SpaceView3D], Optional[Area]): the first 3D view that we find
-            in the given Blender screen and its area, or ``(None, None)`` if
-            no 3D view was found
+        the first 3D view that we find in the given Blender screen and its area,
+        or ``(None, None)`` if no 3D view was found
     """
-    for area in screen.areas:
+    for area in screen.areas:  # type: ignore
         if area.type == "VIEW_3D":
             for space in area.spaces:
                 if space.type == "VIEW_3D":